requirements.txt
python-telegram-bot==20.7
python-dotenv==1.0.0
SQLAlchemy==2.0.23
aiosqlite==0.19.0
greenlet==3.0.3 